backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy
python-dotenv
bcrypt
PyJWT
flask-cors
flask-login
psycopg2-binary